+++ web/app/Models/Project.php
@@ -2,12 +2,13 @@
 
 namespace App\Models;
 
+use App\Models\Concerns\BelongsToAuthenticatedUser;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Project extends Model
 {
-    use HasFactory;
+    use BelongsToAuthenticatedUser, HasFactory;
 
     protected $fillable = [
         'created_by',